board/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
amqp==5.0.6
arrow==1.1.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
binaryornot==0.4.4
blis==0.7.4
bs4==0.0.1
cached-property==1.5.2
catalogue==1.0.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
cymem==2.0.5
DAWG-Python==0.7.2
defusedxml==0.7.1
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-dotenv==1.4.2
django-filter==2.4.0
django-preferences==1.0.0
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docker==5.0.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jinja2-time==0.2.0
kombu==5.1.0
lxml==4.6.3
mailchimp-marketing==3.0.55
MarkupSafe==2.0.1
more-itertools==8.8.0
multidict==5.1.0
murmurhash==1.0.5
numpy==1.21.1
oauthlib==3.1.1
ordered-set==3.1.1
ordered-set-stubs==0.1.3
Pillow==8.3.1
plac==1.1.3
poetry-version==0.1.5
poyo==0.5.0
preshed==3.0.5
profanity-filter==1.3.3
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.100
selenium==3.141.0
six==1.16.0
smart-open==5.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
spacy==2.3.7
sqlparse==0.4.1
srsly==1.0.5
text-unidecode==1.3
thinc==7.4.5
tomlkit==0.5.11
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
views==0.3
vine==5.0.0
wasabi==0.8.2
wcwidth==0.2.5
websocket-client==1.1.0
XlsxWriter==1.4.4
yarl==1.6.3
